pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "math-ml-research-copilot"
version = "0.1.0"
description = "Local-first MVP for ML/math paper ingestion, retrieval, QA and comparison."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "streamlit>=1.48.0",
  "pydantic-settings>=2.10.0",
  "python-multipart>=0.0.20",
  "pymupdf>=1.26.0",
  "faiss-cpu>=1.12.0",
  "numpy>=2.3.0",
  "openai>=1.100.0",
  "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0",
  "ruff>=0.12.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
